solution/service_stg/requirements.txt
asyncio
APScheduler
confluent_kafka
flask
psycopg
psycopg-binary
pydantic
redis
kafka-python